% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_thresholddata.R
\name{extract_thresholddata}
\alias{extract_thresholddata}
\title{A function which extracts tresholdData that was measured with USZ_STT}
\usage{
extract_thresholddata(list_datafiles)
}
\arguments{
\item{list_datafiles}{A list with xls(x) files}
}
\description{
This function allows to read in, transpose and clean auditory
threshold data (xlsx format) from multiple
participants and converts it into a single dataframe
}
\examples{
example_data_path <- system.file("exdata","exdatathr.xls", package = "UZSmultiThreshold")
file_list <- list.files(example_data_path, full.names = TRUE, recursive = TRUE)
Call extract_thresholddata function
threshold_data <- extract_thresholddata(file_list)
head(threshold_data)
}
\keyword{data}
\keyword{extraction,}
\keyword{merging}
